# Replace CreateStreamOnHGlobal with a custom COM Memory Stream