--- a/src/Compilers/Core/Portable/PEWriter/IUnsafeComStream.cs
+++ b/src/Compilers/Core/Portable/PEWriter/IUnsafeComStream.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.ComTypes;
+
+namespace Roslyn.Utilities
+{
+    /// <summary>
+    /// This is a re-definition of COM's IStream interface. The important change is that
+    /// the Write method takes an <see cref="IntPtr"/> instead of a byte[] to avoid the
+    /// allocation cost when called from native code.
+    /// </summary>
+    [Guid("0000000c-0000-0000-C000-000000000046"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown), ComImport]
+    internal interface IUnsafeComStream
+    {
+        // ISequentialStream portion
+        void Read([MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 1), Out] byte[] pv, int cb, IntPtr pcbRead);
+        void Write(IntPtr pv, int cb, IntPtr pcbWritten);
+
+        // IStream portion
+        void Seek(long dlibMove, int dwOrigin, IntPtr plibNewPosition);
+        void SetSize(long libNewSize);
+        void CopyTo(IStream pstm, long cb, IntPtr pcbRead, IntPtr pcbWritten);
+        void Commit(int grfCommitFlags);
+        void Revert();
+        void LockRegion(long libOffset, long cb, int dwLockType);
+        void UnlockRegion(long libOffset, long cb, int dwLockType);
+        void Stat(out STATSTG pstatstg, int grfStatFlag);
+        void Clone(out IStream ppstm);
+    }
+
+}