--- a/src/Compilers/Core/Portable/PEWriter/IUnsafeComStream.cs
+++ b/src/Compilers/Core/Portable/PEWriter/IUnsafeComStream.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+using System;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.ComTypes;