CreateStreamOnHGlobal turns out to have terrible performance for the access pattern used by DiaSymReader (lots of small "SetSize" calls)

So, now we implement the stream in managed code. The main advantage of this implementation is that it uses byte[] chunks rather than trying to keep a contiguous memory block.

Also, we redefine the IStream interface with a small tweak to "Write" to avoid allocating byte[] when called from native code.