@pharring 2015-04-23
@tmat @agocke @jaredpar for review.