@pharring 2015-04-23
The performance improvement is quite impressive (or rather, it shows how slow CreateStreamOnHGlobal was):
C# compiler throughput (Roslyn snapshot):
Before: 41.559 seconds CPU time, 25.010 seconds elapsed
After:30.882 seconds CPU time, 14.519 seconds elapsed