@agocke 2015-04-23
Won't this require the same PInvoke overhead that we saw the first time we stopped using a native stream?