@pharring 2015-04-23
@agocke That's been (somewhat) mitigated by using an IntPtr for Write. Once we get new (official) perf traces, I'll look to see if I can squeeze any more out of it.